requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-functions
google-cloud-pubsub
functions-framework
pandas
python-dotenv
google-generativeai
